detect more syntax errors (issue #24)
unfortunately it's necessary to peek the outputted AST to accomplish this, which may be unreliable. the same kludge is already used in isCallExpression()

the "caller can supply their own AST construction methods" feature probably needs to go. either that or some parsing functions need to return two values: one for the constructed AST and some for internal use - that would be rather wasteful.