Use a different style of escaping in test cases
Using the same backslash escapes to represent characters in source code
as Lua uses to represent them inside string literals is just asking for
a bug in which a backslash is counted a wrong number of times somewhere.
Instead using a letter surrounded by @ characters, as it's visually
distinctive and unlikely to appear in normal source code anyway.

Test cases updated appropriately; expected results didn't change.