Add a hack to make the test suite pass again in PhantomJS
Newer versions of PhantomJS make the .line and .column properties of
SyntaxError objects read-only, and writing to read-only properties in
strict mode throws a TypeError.  Thus, when a syntax error appears in
the input, an error is thrown with a different type and message from
what the test suite expects, which makes test cases fail.

It was a mistake to piggyback on SyntaxError for errors in inputs to
luaparse in the first place. After all, a SyntaxError already has a
meaning: it is thrown by the JavaScript interpreter when it fails to
parse code it is supposed to interpret. Code which attempts to catch
SyntaxErrors from the JS interpreter will be misled by errors
generated by luaparse.

A proper fix will change the error thrown to be a custom type, and no
longer a SyntaxError. This will be a breaking change for some downstream
users; a comment has been left acknowledging the problem.