Add 'LuaJIT' as a supported version, with contextual goto
This enables features present in LuaJIT 2.1.0-beta3 with Lua 5.2
compatibility options disabled. This is the version available in the
Debian unstable repository at the time of this writing.  The main
difference from the other dialects is that the token "goto" is treated
like a keyword if found in statement context and followed by an
identifier, and like an identifier otherwise.

This functionality should be considered unstable; some choices made here
may be revisited and changed in a backwards-incompatible manner.

Things to consider later:

- Future versions of LuaJIT may add more features, making this option's
  behaviour inaccurate.
- LuaJIT can be compiled in Lua 5.2 compatibility mode, which changes
  the set of available features (including contextual goto).
- Decide whether choosing the LuaJIT dialect should automatically
  enable extendedIdentifiers.