Refactor handling of unrecognised escape sequences
This prevents a hypothetical issue where for example '\x' would
be interpreted as '\u' if the unicodeEscapes feature were enabled,
but hexEscapes were not. There is currently no way to trigger this,
but there might have been if individual feature flags were exposed
in the API surface; doing so is under consideration.