Remove docco
Docco version 0.8.1 introduced a regression in
<https://github.com/jashkenas/docco/pull/414>, causing annotated
source to no longer be generated. This was discovered in the course
of CI migration.

Since I am not big on literate programming (in fact, I see it as very
much pointless), instead of debugging it, I decided to scrap docco
entirely. This removed a whole slate of stale development dependencies,
which only makes me happier.