Tweak automatic release CI configuration
Because the 'created' type (configured by default) triggers the
workflow only for immediately-published releases, I managed to skip
the hook by first making a draft release and only then publishing
it.

Thank you, GitHub, for the oh-so-intuitive and not at all full of
pitfalls productivity tool.

Also, remove publishing to the GitHub package registry, since I do
not have that set up.